people details in one slice & saga
