+++ src/features/people/peopleSaga.js
@@ -7,9 +7,13 @@ import {
     selectPage,
     setLoadingState,
     setPeopleQuery,
-    setPeoplePage
+    setPeoplePage,
+    selectId,
+    fetchPersonDetailsSuccess,
+    fetchPersonCreditsSuccess,
+    fetchPersonDetails
 } from "./peopleSlice";
-import { getPeople } from "./api";
+import { getPeople, getPersonCredits, getPersonDetails } from "./api";
 
 function* fetchPeopleHandler() {
     const page = yield select(selectPage);
@@ -29,14 +33,28 @@ function* fetchPeopleHandler() {
     }
 };
 
+function* fetchPersonDetailsHandler() {
+    try {
+      const id = yield select(selectId);
+      const person = yield call(getPersonDetails, id);
+      const credits = yield call(getPersonCredits, id);
+      yield delay(500);
+      yield put(fetchPersonDetailsSuccess(person));
+      yield put(fetchPersonCreditsSuccess(credits));
+    } catch (error) {
+      yield put(fetchPeopleError());
+    }
+  };
+
 function* setOnChangeHandler() {
-    yield put(setLoadingState());
+    yield put(setLoadingState(true));
     yield delay(500);
     yield put(fetchPeople());
 };
 
-export function* watchFetchPopularPeople() {
+export function* peopleSaga() {
     yield takeLatest(setPeopleQuery.type, setOnChangeHandler)
     yield takeLatest(setPeoplePage.type, setOnChangeHandler);
     yield takeLatest(fetchPeople.type, fetchPeopleHandler);
+    yield takeLatest(fetchPersonDetails.type, fetchPersonDetailsHandler);
 }
